requirements.txt
Django~=3.1.2
python-decouple~=3.3
djangorestframework~=3.12.2